Split IVsService and VsService into separate files
