+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/VsService`1.cs
@@ -44,22 +44,4 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS
             return typeof(T);
         }
     }
-
-    /// <summary>
-    ///     Provides an implementation of <see cref="IVsService{TInterfaceType, TServiceType}"/> that calls into Visual Studio's <see cref="SVsServiceProvider"/>.
-    /// </summary>
-    [Export(typeof(IVsService<,>))]
-    internal class VsService<TService, TInterface> : VsService<TInterface>, IVsService<TService, TInterface>
-    {
-        [ImportingConstructor]
-        public VsService([Import(typeof(SVsServiceProvider))]IServiceProvider serviceProvider, IProjectThreadingService threadingService)
-            : base(serviceProvider, threadingService)
-        {
-        }
-
-        protected override Type GetServiceType()
-        {
-            return typeof(TService);
-        }
-    }
 }
